rename mbtiles files
