+++ martin-mbtiles/src/lib.rs
@@ -6,10 +6,10 @@ pub use errors::{MbtError, MbtResult};
 mod mbtiles;
 pub use mbtiles::{IntegrityCheckType, Mbtiles, Metadata};
 
-mod mbtiles_pool;
-pub use mbtiles_pool::MbtilesPool;
+mod pool;
+pub use pool::MbtilesPool;
 
-mod tile_copier;
-pub use tile_copier::{apply_mbtiles_diff, CopyDuplicateMode, TileCopier};
+mod copier;
+pub use copier::{apply_mbtiles_diff, CopyDuplicateMode, MbtilesCopier};
 
-mod mbtiles_queries;
+mod queries;
